backend/pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0",
    "django-allauth[headless,mfa]>=65.13.1",
    "django-cors-headers>=4.9.0",
    "fido2>=2.0.0",
    "psycopg2-binary>=2.9",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.30.0",
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
    "djangorestframework>=3.14.0",
    "openpyxl>=3.1.0",
]

